openssl: Removed -DHAVE_CRYPTODEV
This is a kernel specific features and unfortunately not all kernels
includes this feature. Some kernels or kernel configurations do not
include include/cryptodev.h.

In case openssl is configured to use the kernel crypto device for hw
accelerated encryption then it will have to be added and overlayed for
that specific board configuration.