systemd: Disabled networkd and pager
Since we are using connman for network manger we dont need
systemd-networkd. Also disabled systemds very annoying use of pager for
output because it simply messes up the output.

Arrghh, why do systemd even try to create a new network manager!? - they
should team up with the connman team and tightly integrate support for
connman instead because it will take forever for systemd-networkd to
reach the feature level of connman... what a waste.